Refactor adding chakra-components pagination components (and other fixes)
