+++ src/components/Accounts/Detail.tsx
@@ -2,16 +2,16 @@ import { Tab, TabList, TabPanel, TabPanels, VStack } from '@chakra-ui/react'
 import { OrganizationHeader, OrganizationName } from '@vocdoni/chakra-components'
 import { useOrganization } from '@vocdoni/react-providers'
 import { Trans, useTranslation } from 'react-i18next'
+import AccountFees from '~components/Accounts/Details/Fees'
+import AccountTransfers from '~components/Accounts/Details/Transfers'
 import { ReducedTextAndCopy } from '~components/Layout/CopyButton'
 import { HeroHeaderLayout } from '~components/Layout/HeroHeaderLayout'
 import { RouteParamsTabs } from '~components/Layout/RouteParamsTabs'
 import { RawContentBox } from '~components/Layout/ShowRawButton'
-import { FallbackHeaderImg, RoutePath } from '~constants'
-import AccountTransfers from '~components/Accounts/Details/Transfers'
-import AccountElections from './Details/Elections'
-import AccountDetails from './Details/AccountDetails'
-import AccountFees from '~components/Accounts/Details/Fees'
 import TextAndTag from '~components/Layout/TextAndTag'
+import { FallbackHeaderImg, RoutePath } from '~constants'
+import AccountDetails from './Details/AccountDetails'
+import AccountElections from './Details/Elections'
 
 const AccountDetail = () => {
   const { organization } = useOrganization()
