+++ src/components/Accounts/Details/AccountDetails.tsx
@@ -1,11 +1,11 @@
 import { Flex, Icon, Link } from '@chakra-ui/react'
 import { OrganizationDescription } from '@vocdoni/chakra-components'
+import { useOrganization } from '@vocdoni/react-providers'
 import { ensure0x } from '@vocdoni/sdk'
 import { Trans, useTranslation } from 'react-i18next'
 import { FaUserAlt } from 'react-icons/fa'
 import { DetailsGrid, GridItemProps } from '~components/Layout/DetailsGrid'
 import { AppBaseURL } from '~constants'
-import { useOrganization } from '@vocdoni/react-providers'
 
 const AccountDetails = () => {
   const { t } = useTranslation()
