+++ src/components/Accounts/Details/Transfers.tsx
@@ -1,5 +1,5 @@
 import { Box, Flex, Icon, IconProps, Link, Skeleton, Td, Text, Th, Thead, Tooltip, Tr } from '@chakra-ui/react'
-import { useOrganization } from '@vocdoni/react-providers'
+import { PaginationProvider, useOrganization, usePagination } from '@vocdoni/react-providers'
 import { AccountData, ITransfer } from '@vocdoni/sdk'
 import { formatDistance } from 'date-fns'
 import { Trans, useTranslation } from 'react-i18next'
@@ -7,9 +7,7 @@ import { BiLogInCircle, BiLogOutCircle } from 'react-icons/bi'
 import { generatePath, Link as RouterLink } from 'react-router-dom'
 import { PaginatedAsyncTable } from '~components/Layout/AsyncList'
 import { ReducedTextAndCopy } from '~components/Layout/CopyButton'
-import { PaginationProvider, usePagination } from '~components/Pagination/PaginationProvider'
 import { RoutePath } from '~constants'
-import { useDateFns } from '~i18n/use-date-fns'
 import { useAccountTransfers } from '~queries/accounts'
 import { generateListStub, PaginationStub } from '~utils/stubs'
 
@@ -34,29 +32,23 @@ const FromToIcon = ({ isIncoming, ...rest }: { isIncoming: boolean } & IconProps
   )
 }
 
-const AccountTransfers = () => {
-  return (
-    <PaginationProvider>
-      <AccountTransfersTable />
-    </PaginationProvider>
-  )
-}
+const AccountTransfers = () => (
+  <PaginationProvider>
+    <AccountTransfersTable />
+  </PaginationProvider>
+)
 
 const AccountTransfersTable = () => {
   const { page } = usePagination()
-  const { formatDistance } = useDateFns()
-  const { t } = useTranslation()
   const { organization } = useOrganization()
 
-  if (!organization) return null
-
-  const txCount = organization.transfersCount ?? 0
+  const txCount = organization?.transfersCount ?? 0
 
   const { data, isError, error, isPlaceholderData } = useAccountTransfers({
-    address: organization.address,
+    address: organization?.address as string,
     page: page,
     options: {
-      enabled: txCount > 0,
+      enabled: !!organization && txCount > 0,
       placeholderData: {
         transfers: generateListStub<ITransfer>({
           amount: 12,
@@ -70,6 +62,8 @@ const AccountTransfersTable = () => {
       },
     },
   })
+
+  if (!organization) return null
 
   return (
     <PaginatedAsyncTable
