+++ src/components/Envelope/EnvelopeList.tsx
@@ -1,14 +1,13 @@
 import { Flex, Grid } from '@chakra-ui/react'
-import { useElection } from '@vocdoni/react-providers'
+import { Pagination } from '@vocdoni/chakra-components'
+import { PaginationProvider, useElection, usePagination } from '@vocdoni/react-providers'
 import { PublishedElection, VoteSummary } from '@vocdoni/sdk'
 import { useTranslation } from 'react-i18next'
 import { ListDataDisplay } from '~components/Layout/AsyncList'
 import { NoResultsError } from '~components/Layout/ContentError'
-import { Pagination } from '~components/Pagination/Pagination'
-import { PaginationProvider, usePagination } from '~components/Pagination/PaginationProvider'
 import { useElectionVotesList } from '~queries/processes'
-import { EnvelopeCard } from './EnvelopeCard'
 import { generateListStub, PaginationStub } from '~utils/stubs'
+import { EnvelopeCard } from './EnvelopeCard'
 
 export const PaginatedEnvelopeList = () => {
   const { election: e } = useElection()
