+++ src/components/Verify/index.tsx
@@ -1,10 +1,10 @@
-import addVote from '/images/add-vote.svg'
 import { Box, Button, Flex, Image, Text } from '@chakra-ui/react'
 import { useState } from 'react'
 import { Trans, useTranslation } from 'react-i18next'
 import { generatePath, useNavigate, useParams } from 'react-router-dom'
 import { InputSearch } from '~components/Layout/Inputs'
 import { RoutePath } from '~constants'
+import addVote from '/images/add-vote.svg'
 
 const SearchVote = ({ compact }: { compact?: boolean }) => {
   const { verifier: urlVerifier }: { verifier?: string } = useParams()
