+++ src/main.tsx
@@ -2,10 +2,10 @@ import { ChakraProvider } from '@chakra-ui/react'
 import { StrictMode } from 'react'
 import * as ReactDOM from 'react-dom/client'
 import { App } from './App'
+import './i18n'
 import reportWebVitals from './reportWebVitals'
 import * as serviceWorker from './serviceWorker'
 import theme from './theme'
-import './i18n'
 
 const container = document.getElementById('root')
 if (!container) throw new Error('Failed to find the root element')
