+++ src/pages/Home.tsx
@@ -1,6 +1,6 @@
-import LandingPage from '~components/Home'
-import { useLoaderData } from 'react-router-dom'
 import { IChainGetInfoResponse } from '@vocdoni/sdk'
+import { useLoaderData } from 'react-router-dom'
+import LandingPage from '~components/Home'
 import { useChainInfo } from '~queries/stats'
 
 const Home = () => {
