+++ src/pages/block.tsx
@@ -1,6 +1,6 @@
-import { BlockDetail as BlockDetailView } from '~components/Blocks/Detail'
-import { useLoaderData } from 'react-router-dom'
 import { IChainBlockInfoResponse } from '@vocdoni/sdk'
+import { useLoaderData } from 'react-router-dom'
+import { BlockDetail as BlockDetailView } from '~components/Blocks/Detail'
 
 const BlockDetail = () => {
   const block = useLoaderData() as IChainBlockInfoResponse
