+++ src/pages/blocks.tsx
@@ -1,7 +1,7 @@
-import ListPageLayout from '~components/Layout/ListPageLayout'
 import { useTranslation } from 'react-i18next'
-import { useChainInfo } from '~queries/stats'
 import { BlocksFilter, PaginatedBlocksList } from '~components/Blocks/BlocksList'
+import ListPageLayout from '~components/Layout/ListPageLayout'
+import { useChainInfo } from '~queries/stats'
 
 const BlocksList = () => {
   const { t } = useTranslation()
