+++ src/pages/transaction.tsx
@@ -1,6 +1,6 @@
+import { ErrTransactionNotFound, Tx } from '@vocdoni/sdk'
 import { useLoaderData } from 'react-router-dom'
 import { TransactionDetail as TransactionDetailView } from '~components/Transactions/Detail'
-import { ErrTransactionNotFound, Tx } from '@vocdoni/sdk'
 
 const TransactionDetail = () => {
   const tx = useLoaderData() as Tx
