+++ src/pages/verify.tsx
@@ -1,11 +1,11 @@
-import { useParams } from 'react-router-dom'
-import { useVoteInfo } from '~queries/vote'
-import { VerifyForm, VerifyFormMinified } from '~components/Verify'
-import { Loading } from '~components/Layout/Loading'
 import { Flex, Heading } from '@chakra-ui/react'
 import { Trans } from 'react-i18next'
+import { useParams } from 'react-router-dom'
 import EnvelopeDetail from '~components/Envelope/Detail'
+import { Loading } from '~components/Layout/Loading'
+import { VerifyForm, VerifyFormMinified } from '~components/Verify'
 import { RoutePath } from '~constants'
+import { useVoteInfo } from '~queries/vote'
 
 const Verify = () => {
   const { verifier }: { verifier?: string } = useParams()
