+++ src/theme/components/Icons.ts
@@ -1,10 +1,10 @@
 import { IconType } from 'react-icons'
 import { BiTransferAlt } from 'react-icons/bi'
-import { HiOutlineCube } from 'react-icons/hi2'
-import { IoTimeOutline } from 'react-icons/io5'
-import { GrValidate } from 'react-icons/gr'
 import { FaExternalLinkAlt } from 'react-icons/fa'
 import { FiInfo } from 'react-icons/fi'
+import { GrValidate } from 'react-icons/gr'
+import { HiOutlineCube } from 'react-icons/hi2'
+import { IoTimeOutline } from 'react-icons/io5'
 
 type IconName = 'BlockIcon' | 'TxIcon' | 'ClockIcon' | 'ValidatorIcon' | 'ExternalIcon' | 'InfoIcon'
