+++ src/theme/components/index.ts
@@ -1,3 +1,5 @@
+import Envelope from '~src/theme/components/Envelope'
+import { QuestionsTypeBadge } from '~src/theme/components/QuestionsTypeBadge'
 import Card from './Card'
 import { ElectionHeader, ElectionTitle } from './Election'
 import ElectionResults from './ElectionResults'
@@ -5,8 +7,6 @@ import Link from './Link'
 import Questions from './Questions'
 import Radio from './Radio'
 import Tag from './Tag'
-import Envelope from '~src/theme/components/Envelope'
-import { QuestionsTypeBadge } from '~src/theme/components/QuestionsTypeBadge'
 
 const components = {
   Card,
